Refactor PeakCoreData folder structure
