+++ PeakCoreData/Core/Fetched Data Sources/FetchedDataProvider.swift
@@ -7,7 +7,6 @@
 //  Copyright © 2018 3Squared Ltd. All rights reserved.
 //
 
-import UIKit
 import CoreData
 
 public enum FetchedUpdate<Object> {
@@ -20,87 +19,11 @@ public enum FetchedUpdate<Object> {
 }
 
 public protocol HasEmptyView: class {
-    var emptyView: UIView? { get }
+    var emptyView: PeakView? { get }
 }
 
 public extension HasEmptyView {
-    var emptyView: UIView? { return nil }
-}
-
-public protocol TableViewUpdatable: class {
-    associatedtype Object: NSManagedObject
-    associatedtype Cell: UITableViewCell
-    func configure(_ cell: Cell, with object: Object)
-}
-
-extension TableViewUpdatable {
-    
-    public func process(updates: [FetchedUpdate<Object>], for tableView: UITableView, with animation: UITableView.RowAnimation = .automatic, completion: ((Bool) -> Void)? = nil) {
-        let batchUpdates: () -> Void = { [weak self] in
-            guard let strongSelf = self else { return }
-
-            updates.forEach { (update) in
-                switch update {
-                case .insert(let indexPath):
-                    tableView.insertRows(at: [indexPath], with: animation)
-                case .update(let indexPath, let object):
-                    guard let cell = tableView.cellForRow(at: indexPath) as? Cell else { return }
-                    strongSelf.configure(cell, with: object)
-                case .move(let indexPath, let newIndexPath):
-                    tableView.moveRow(at: indexPath, to: newIndexPath)
-                case .delete(let indexPath):
-                    tableView.deleteRows(at: [indexPath], with: animation)
-                case .deleteSection(let section):
-                    tableView.deleteSections(IndexSet(integer: section), with: animation)
-                case .insertSection(let section):
-                    tableView.insertSections(IndexSet(integer: section), with: animation)
-                }
-            }
-        }
-        
-        if #available(iOS 11.0, *) {
-            tableView.performBatchUpdates(batchUpdates, completion: completion)
-        } else {
-            tableView.beginUpdates()
-            batchUpdates()
-            tableView.endUpdates()
-            completion?(true)
-        }
-    }
-}
-
-public protocol CollectionViewUpdatable: class {
-    associatedtype Object: NSManagedObject
-    associatedtype Cell: UICollectionViewCell
-    func configure(_ cell: Cell, with object: Object)
-}
-
-extension CollectionViewUpdatable {
-    
-    public func process(updates: [FetchedUpdate<Object>], for collectionView: UICollectionView, completion: ((Bool) -> Void)? = nil) {
-        let batchUpdates: () -> Void = { [weak self] in
-            guard let strongSelf = self else { return }
-            
-            updates.forEach { (update) in
-                switch update {
-                case .insert(let indexPath):
-                    collectionView.insertItems(at: [indexPath])
-                case .update(let indexPath, let object):
-                    guard let cell = collectionView.cellForItem(at: indexPath) as? Cell else { return }
-                    strongSelf.configure(cell, with: object)
-                case .move(let indexPath, let newIndexPath):
-                    collectionView.moveItem(at: indexPath, to: newIndexPath)
-                case .delete(let indexPath):
-                    collectionView.deleteItems(at: [indexPath])
-                case .deleteSection(let section):
-                    collectionView.deleteSections(IndexSet(integer: section))
-                case .insertSection(let section):
-                    collectionView.insertSections(IndexSet(integer: section))
-                }
-            }
-        }
-        collectionView.performBatchUpdates(batchUpdates, completion: completion)
-    }
+    var emptyView: PeakView? { return nil }
 }
 
 protocol FetchedDataProviderDelegate: class {
