+++ PeakCoreData/Core/Protocols/ManagedObjectType.swift
